requirements.txt
google-cloud-storage==2.17.0
joblib==1.4.2